Split up and rename some components
