+++ packages/ndla-ui/src/Frontpage/index.ts
@@ -1,7 +1,9 @@
-export { default as FrontpageHeaderNew } from './FrontpageHeaderNew';
+export { default as FrontpageHeader } from './FrontpageHeader';
 export { default as FrontpageInfo } from './FrontpageInfo';
 export { default as FrontpageFilm } from './FrontpageFilm';
-export { default as FrontpageSubjectIllustration } from './illustrations/FrontpageSubjectIllustration';
+export {
+  default as FrontpageSubjectIllustration,
+} from './illustrations/FrontpageSubjectIllustration';
 export {
   default as FrontpageSubjectsInPortal,
 } from './FrontpageSubjectsInPortal';
